test(java): move Fory JSON concurrency coverage
